Added /sw setvertex, /sw rebackup and ChunkBackup class to restore blocks. Use Player->setImmobile() to disallow player movements. Fixed bug with PvP not cancelling during no-pvp event and fixed return value of SkyWars::canModifyTerrain().
ChunkBackup helps restore the previous level state (blocks and tiles for now)
and deletes all unsavable entities (all entities except Player).

ChunkBackup does NOT create a disk backup of the level folder, it creates
backup of specific chunks and tiles that fall into the two vertexes that you
specify while configuring a new SkyWars game. The backup is CACHED and not
saved to disk.